requirements.txt
google-generativeai
python-decouple
python-dotenv